FINAL CHECK: Added security check for order and orderdetails. Added some styling fixed some errors.

+++ server/api/orderDetails.js
@@ -4,8 +4,12 @@ const {OrderDetails, Order, Product} = require('../db/models')
 //GET ALL ORDER DETAILS
 router.get('/', async (req, res, next) => {
   try {
-    const orderDetails = await OrderDetails.findAll()
-    res.status(200).json(orderDetails)
+    if (req.user) {
+      const orderDetails = await OrderDetails.findAll()
+      res.status(200).json(orderDetails)
+    } else {
+      res.status(401).json('User does not have get orderDetails access.')
+    }
   } catch (err) {
     next(err)
   }
@@ -13,15 +17,19 @@ router.get('/', async (req, res, next) => {
 //GETTING USER ID & SPECIFIED ORDER ID
 router.get('/:userId/', async (req, res, next) => {
   try {
-    const orders = await Order.findOne({
-      where: {
-        userId: req.params.userId,
-        isCompleted: false
-      },
-      include: [Product]
-    })
-    const data = await orders.getProducts()
-    res.json(orders)
+    if (req.user) {
+      const orders = await Order.findOne({
+        where: {
+          userId: req.params.userId,
+          isCompleted: false
+        },
+        include: [Product]
+      })
+      const data = await orders.getProducts()
+      res.json(orders)
+    } else {
+      res.status(401).json('User does not have get orderDetails access.')
+    }
   } catch (err) {
     next(err)
   }
@@ -29,21 +37,25 @@ router.get('/:userId/', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    let order = await Order.findOrCreate({
-      where: {
-        userId: req.body.userId,
-        isCompleted: false
-      },
-      include: [Product]
-    })
-    const product = await Product.findByPk(req.body.productId)
+    if (req.user) {
+      let order = await Order.findOrCreate({
+        where: {
+          userId: req.body.userId,
+          isCompleted: false
+        },
+        include: [Product]
+      })
+      const product = await Product.findByPk(req.body.productId)
 
-    await product.addOrder(order[0], {
-      through: {
-        productQuantity: req.body.productQuantity,
-        productPrice: req.body.productPrice
-      }
-    })
+      await product.addOrder(order[0], {
+        through: {
+          productQuantity: req.body.productQuantity,
+          productPrice: req.body.productPrice
+        }
+      })
+    } else {
+      res.status(401).json('User does not have post orderDetails access.')
+    }
   } catch (err) {
     next(err)
   }
@@ -51,15 +63,19 @@ router.post('/', async (req, res, next) => {
 
 router.put('/:userId/', async (req, res, next) => {
   try {
-    const modifyOrder = await OrderDetails.findOne({
-      where: {
-        userId: req.params.userId,
-        isCompleted: false
-      },
-      include: [Product]
-    })
-    const data = await modifyOrder.getProducts()
-    res.send(await modifyOrder.update(req.body))
+    if (req.user) {
+      const modifyOrder = await OrderDetails.findOne({
+        where: {
+          userId: req.params.userId,
+          isCompleted: false
+        },
+        include: [Product]
+      })
+      const data = await modifyOrder.getProducts()
+      res.send(await modifyOrder.update(req.body))
+    } else {
+      res.status(401).json('User does not have put orderDetails access.')
+    }
   } catch (err) {
     next(err)
   }
@@ -67,13 +83,17 @@ router.put('/:userId/', async (req, res, next) => {
 
 router.delete('/', async (req, res, next) => {
   try {
-    let result = await OrderDetails.destroy({
-      where: {
-        orderId: req.body.orderId,
-        productId: req.body.productId
-      }
-    })
-    res.json(result)
+    if (req.user) {
+      let result = await OrderDetails.destroy({
+        where: {
+          orderId: req.body.orderId,
+          productId: req.body.productId
+        }
+      })
+      res.json(result)
+    } else {
+      res.status(401).json('User does not have delete orderDetails access.')
+    }
   } catch (err) {
     next(err)
   }
